Deleted comparefonts module, moved contents to models
