+++ app/main.py
@@ -8,7 +8,6 @@ from rethinkdb.errors import RqlRuntimeError, RqlDriverError
 
 import downloadfonts
 import models
-from comparefonts import compare_fonts
 from glyphpalette import fonts_all_glyphs
 import init_db
 from utils import filename_to_family_name
@@ -85,7 +84,7 @@ def upload_fonts():
     fontset = models.add_fontset(fonts_before, fonts_after, uuid)
     r.table('fontsets').insert(fontset).run(g.rdb_conn)
 
-    comparison = compare_fonts(fonts_before, fonts_after, uuid)
+    comparison = models.add_fonts_comparison(fonts_before, fonts_after, uuid)
     r.table('comparisons').insert(comparison).run(g.rdb_conn)
     fonts_glyphsets = fonts_all_glyphs(fonts_before, fonts_after, uuid)
     r.table('glyphs').insert(fonts_glyphsets).run(g.rdb_conn)
@@ -168,7 +167,7 @@ def api_upload_fonts(upload_type):
         fontset = models.add_fontset(fonts_before, fonts_after, uuid)
         r.table('fontsets').insert(fontset).run(g.rdb_conn)
 
-        comparison = compare_fonts(fonts_before, fonts_after, uuid)
+        comparison = add_fonts_comparison(fonts_before, fonts_after, uuid)
         r.table('comparisons').insert(comparison).run(g.rdb_conn)
 
         fonts_glyphsets = fonts_all_glyphs(fonts_before, fonts_after, uuid)
